backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.23
pymysql
alembic
python-dotenv
pydantic
pytz
apscheduler
python-multipart
requests
